\name{Lorber-Egeghy-East Model Concentration (LEEM Concentration)}
\alias{LEEM_Concentration}
\title{Concentration Summary Statistics to Weighted Concentration Estimates.}
\description{
A function which creates synthetic lognormal data from a dataframe of summary statistic concentration data, a selected weighting column, n number of points.

This model is based on methods described in Egeghy & Lorber (2011), Lorber & Egeghy (2011) , and East et al. (2021):

              Egeghy & Lorber: https://pubmed.ncbi.nlm.nih.gov/20145679/
              Lorber & Egeghy: https://pubmed.ncbi.nlm.nih.gov/21517063/
              East et al.    : https://www.nature.com/articles/s41370-021-00374-w

This model uses a series of estimating equations for geometric mean (GM) and
geometric standard deviation (GSD) primarily adapted from Pleil et al. (2014).

              Pleil et al.   : https://pubmed.ncbi.nlm.nih.gov/25333994/


Accepted Units: "ng/m3", "ug/m3", "pg/ml", "ug/L", "ng/L", "pg/g", "ug/kg", "ng/g"


Concentration output: "ng/m3", "ng/L", or "ng/g"
Exposure  output: "ng/day"


The function returns a list of dataframes:

Summary: Mean, min, 10th percentile, median, 75th percentile, 95th percentile, and max for concentration across each media, pathway, and chemical entered.

Data: The input data for which GM and GSDs could be estimated.

Not Used: The input data for which GM and GSDS could not be estimated.

Used Row Count: Counts of used datasets by chemical and media.

Raw: Raw data for plotting.

Metadata: Time and date, chemicals, media, n, and seed used in the run.


}
\usage{
LEEM_Concentration(data,wtcol,n,seed = NULL)

This function is used when concentration data is to be summarized but there are no exposure factors or individuals to apply. This is done for PFOA in serum in the example below but can be applied to other media such as wastewater and effluent.
}
\arguments{

data = A dataframe containing the following columns: "Chemical", "Media","Units","Sample Size","Min","Max","Median","Mean","SD","GM","GSD","P10","P25","P75","P90","P95","P99." Column names are not case sensitive.

wtcol = A character input for the name of the 'weighting' column (often 'sample size')

n = the number of generated concentration and exposure outcomes generated by rlnorm().

seed = Optional input to ensure or characterize variability between runs. Set by default to 12345.


}

\examples{
# Using the Example dataset "LEEMR_Example_Data":

LEEM(LEEMR_Example_Data$Serum,
     "sample size",
     1000)
}

