\name{Lorber Egeghy Model}
\alias{LEM_Concentration}
\title{Concentration Summary Statistics to Weighted Concentration Curves.}
\description{
A function which creates synthetic lognormal data from a dataframe of summary statistic concentration data,a selected weighting column, and n number of points.

This model is based on methods described in Egeghy & Lorber:
              "An assessment of the exposure of Americans to perfluorooctane
              sulfonate: a comparison of estimated intake with values inferred from
              NHANES data"(https://pubmed.ncbi.nlm.nih.gov/20145679/) and Lorber & Egeghy:
              Simple intake and pharmacokinetic modeling to characterize exposure of Americans to
              perfluoroctanoic acid, PFOA(https://pubmed.ncbi.nlm.nih.gov/21517063/)

The methods described above have been modified to include a geometric mean (GM) and
geometric standard deviation (GSD) estimating sequence from Pleil et. al (https://pubmed.ncbi.nlm.nih.gov/25333994/).

Accepted Units: "ng/m3", "ug/m3", "pg/ml", "ug/L", "ng/L", "pg/g", "ug/kg", "ng/g"


Concentration output: "ng/m3", "ng/L", or "ng/g"


WGMs and WGSDs are determined by complete cases of both, as visible in the 'Dataframe' in the output list.

Function output returns three objects:

Summary Statistics: the result, mean, min, 10th percentile, median, 75th percentile, 95th percentile, max for concentration.

Concentration: The generated lognormal concentration curve with "n" number of points.

Dataframe: The input data used (df), including estimated summary statistics.

}
\usage{
LEM_Concentration(df,wtcol,n,)

The same seed (12345) is set for every run.
}
\arguments{

df = A dataframe containing the following columns:
"Units","Sample Size","Min","Max","Median","Mean","SD","GM","GSD","P10","P25","P75","P90","P95","P99"

wtcol = A character input for the name of the 'weighting' column (often sample size)

n = the number of generated exposure outcomes generated by rlnorm().




}

\examples{
# Assuming a user has assembled an input "mydata" as described above:
LEM_Concentration(mydata,"Sample Size",10000)
}
