\name{Lorber Egeghy Model}
\alias{LEM_Template}
\title{Formatting for the input to the Lorber Egeghy Model.}
\description{
A function which creates an Excel file with the required column names to run the full Lorber Egeghy Model.
Two sheets are created: one for studies with concentration summary statistics, and another for factors to be assigned to the generated lognormal curves (normally exposure factors).
}
\usage{
LEM_Template()

No arguments are passed to LEM_Template. An Excel file is generated in the user's working directory.
}
\arguments{
None.


}
