\name{Lorber-Egeghy-East Model (LEEM)}
\alias{LEEM}
\title{Concentration Summary Statistics to Weighted Concentration & Exposure Estimates.}
\description{
A function which creates synthetic lognormal exposure data from a dataframe of summary statistic concentration data, exposure factors, a selected weighting column, n number of points.

This model is based on methods described in \href{https://pubmed.ncbi.nlm.nih.gov/20145679/}{Egeghy & Lorber (2011)} , \href{https://pubmed.ncbi.nlm.nih.gov/21517063/}{Lorber & Egeghy (2011)} , and \href{https://www.nature.com/articles/s41370-021-00374-w}{East et al. (2021)}.

The series of estimating equations for geometric mean (GM) and
geometric standard deviation (GSD) is primarily adapted from \href{https://pubmed.ncbi.nlm.nih.gov/25333994/}{Pleil et al. (2014)}.




\describe{
    \item{Accepted Units: \emph{ng/m3}, \emph{ug/m3}, \emph{pg/ml}, \emph{ug/L}, \emph{ng/L},\emph{pg/g},\emph{ug/kg}, \emph{ng/g}
    }
    \item{ }
    \item{Concentration output: \emph{ng/m3}, \emph{ng/L}, or \emph{ng/g}}
    \item{Exposure  output: \emph{ng/day}}
}






The function returns a list of dataframes:

Summary: Mean, min, 10th percentile, median, 75th percentile, 95th percentile, and max for concentration and exposure across each media, pathway, and chemical entered.

Data: The input data for which GM and GSDs could be estimated.

Not Used: The input data for which GM and GSDS could not be estimated.

Used Row Count: Counts of used datasets by chemical and media.

Raw: Raw data for plotting.

Metadata: Time and date, chemicals, media, n, and seed used in the run.


}
\usage{
LEEM(data,factors,wtcol,n,seed = NULL)

The LEEM is used to generate exposure estimates from sparse summary statistics
for any number of chemicals, individuals, and media entered in the input file.
Note: \code{\link{LEEM_Template}} generates an Excel file template with for \emph{data} and \emph{factors} arguments.
}
\arguments{

data = A dataframe containing the following columns: \emph{Chemical}, \emph{Media}, \emph{Units}, \emph{Sample Size}, \emph{Min}, \emph{Max}, \emph{Median}, \emph{Mean}, \emph{SD}, \emph{GM}, \emph{GSD}, \emph{P10}, \emph{P25},  \emph{P75}, \emph{P90}, \emph{P95}, \emph{P99.} Column names are not case sensitive.


factors = A dataframe with exposure factors for each media in the \emph{data} dataframe. Contains the following columns : \emph{Path},\emph{Media},\emph{Individual},\emph{Factor}, which are not case sensitive.


wtcol = A character input for the name of the 'weighting' column (often 'sample size')

n = the number of generated concentration and exposure outcomes generated by rlnorm().

seed = Optional input to ensure or characterize variability between runs. Set by default to 12345.


}

\examples{
# Using the Example dataset 'LEEMR_Example_Data':
# Using the Example factors in 'LEEMR_ExampleData""
# Specifying the column "sample size" to weight datasets
# setting n = 1000
# No argument provided for 'seed', defaulting to "12345"

LEEM(LEEMR_Example_Data$Water,
     LEEMR_Example_Data$`Water Factors`,
     "sample size",
     1000)
}
