\name{Lorber-Egeghy-East Model Concentration (LEEM Concentration)}
\alias{LEEM_Concentration}
\title{Concentration Summary Statistics to Weighted Concentration Estimates}
\description{
A function which creates synthetic lognormal data from a dataframe of summary statistic concentration data, exposure factors, a selected weighting column, and \emph{n} number of points for a single media and single chemical.

This model is based on methods described in \href{https://pubmed.ncbi.nlm.nih.gov/20145679/}{Egeghy & Lorber (2011)} , \href{https://pubmed.ncbi.nlm.nih.gov/21517063/}{Lorber & Egeghy (2011)} , and \href{https://www.nature.com/articles/s41370-021-00374-w}{East et al. (2021)}.

The series of estimating equations for geometric mean (GM) and
geometric standard deviation (GSD) is primarily adapted from \href{https://pubmed.ncbi.nlm.nih.gov/25333994/}{Pleil et al. (2014)}.




\describe{
    \item{Accepted Units: \emph{ng/m3}, \emph{ug/m3}, \emph{pg/ml}, \emph{ng/mL}, \emph{ug/L}, \emph{ng/L}, \emph{pg/g}, \emph{ug/kg}, \emph{ng/g}}
    \item{ }
    \item{Concentration output: \emph{ng/m3}, \emph{ng/L}, or \emph{ng/g}}
   }

The function returns a list of dataframes:

Summary: mean, min, 10th percentile, median, 75th percentile, 95th percentile, and max for concentration and exposure across each media, pathway, and chemical entered.

Used Input: The input data for which GM and GSDs could be estimated.

Excluded Input: The input data for which GM and GSDs could not be estimated.

Used Dataset Counts: Counts of used datasets by chemical and media.

Raw: Raw data for plotting.

Metadata: Time and date, chemicals, media, n, and seed used in the run.

}
\usage{
LEEM_Concentration(data, wtcol, n, seed = NULL)

This function is used to summarize lognormally distributed concentration data in similar media.
An example using PFOA in serum is provided in the \emph{Examples} section below.
}

\arguments{

data = A dataframe of one media containing the following columns: \emph{Chemical}, \emph{Media}, \emph{Units}, \emph{Sample Size}, \emph{Min}, \emph{Max}, \emph{Median}, \emph{Mean}, \emph{SD}, \emph{GM}, \emph{GSD}, \emph{P10}, \emph{P25}, \emph{P75}, \emph{P90}, \emph{P95}, \emph{P99.} Column names are not case sensitive.


wtcol = A character input for the name of the 'weighting' column (often 'sample size').

n = The number of generated concentration and exposure outcomes generated by rlnorm().

seed = Optional input to ensure or characterize variability between runs. Set by default to 12345.


}

\examples{
# using the Example dataset "LEEMR_Example_Data":
# dataframe "serum" used as data argument
# specifying the column "sample size" to weight datasets
# setting n = 1000
# no argument provided for 'seed', defaulting to "12345"

LEEM_Concentration(LEEMR_Example_Data$Serum,
                   "sample size",
                   1000)
}

