\name{LEM}
\alias{Lorber Egeghy Model}
\title{Concentration Summary Statistics to Estimated Exposure Curves.}
\usage{
LEM(df,ef,n)
}
\description{
A function which draws exposure curves from a dataframe of summary statistic concentration data, an exposure factor, and n number of points.
}
\examples{
LEM(df,ef,n)
}
