\name{Lorber Egeghy Model}
\alias{LEM}
\title{Concentration Summary Statistics to Estimated Exposure Curves.}
\description{
A function which draws exposure curves from a dataframe of summary statistic concentration data,
a selected weight column, an exposure factor, and n number of points.

This model is based on methods described in Egeghy & Lorber:
              "An assessment of the exposure of Americans to perfluorooctane
              sulfonate: a comparison of estimated intake with values inferred from
              NHANES data"(https://pubmed.ncbi.nlm.nih.gov/20145679/) and Lorber & Egeghy:
              Simple intake and pharmacokinetic modeling to characterize exposure of Americans to
              perfluoroctanoic acid, PFOA(https://pubmed.ncbi.nlm.nih.gov/21517063/)

The methods described above have been modified to include a geometric mean (GM) and
geometric standard deviation (GSD) estimating sequence from Pleil et. al (https://pubmed.ncbi.nlm.nih.gov/25333994/).

Accepted Units: "ng/m3", "ug/m3", "pg/ml", "ug/L", "ng/L", "pg/g", "ug/kg", "ng/g"


Concentration output: "ng/m3", "ng/L", or "ng/g"

}
\usage{
LEM(df,wt,expf,n,seed)

In events where only concentration curves are required, please enter the concentration unit and set expf to 1.
}
\arguments{

df = A dataframe containing the following columns:
"Units","Sample Size","Min","Max","Median","Mean","SD","GM","GSD","P10","P25","P75","P90","P95","P99"

wt = A character input for the name of the 'weight' column (often sample size)

expf = the fraction by which the concentration is translated into exposure.

expf = the units of the exposure. Concentration data is ng/L, ng/m3, or ng/g, which dicates the expf.

n = the number of generated exposure outcomes generated by rlnorm().

seed = a seed used to dictate the random number generation (see "set.seed" for more)


}

\examples{
LEM(df,wt,expf,n,seed)
LEM(data,"Sample Size",1.5,200,12345)
}
