\name{Lorber Egeghy Model}
\alias{LEM}
\title{Concentration Summary Statistics to Estimated Exposure Curves.}
\description{
A function which draws exposure curves from a dataframe of summary statistic concentration data,
a selected weight column, an exposure factor, and n number of points.

This model is based on methods described in Egeghy & Lorber:
              "An assessment of the exposure of Americans to perfluorooctane
              sulfonate: a comparison of estimated intake with values inferred from
              NHANES data"(https://pubmed.ncbi.nlm.nih.gov/20145679/) and Lorber & Egeghy:
              Simple intake and pharmacokinetic modeling to characterize exposure of Americans to
              perfluoroctanoic acid, PFOA(https://pubmed.ncbi.nlm.nih.gov/21517063/)

The methods described above have been modified to include a geometric mean (GM) and
geometric standard deviation (GSD) estimating sequence from Pleil et. al (https://pubmed.ncbi.nlm.nih.gov/25333994/).
}
\usage{
LEM(df,wt,ef,n)
}
\arguments{

df = A dataframe containing the following columns:
"Sample_Size","Min","Max","Median","Mean","SD","GM","GSD","P10","P25","P75","P90","P95","P99"

wt = (not required) A character input for the name of the 'weight' column (often sample size)
     defaults to unweighted avg. GM and GSD.

ef = the fraction by which the concentration is translated into exposure.

n = the number of generated exposure outcomes generated by rlnorm().


}


\examples{
LEM(df,wt,ef,n)
}
