\name{Lorber Egeghy Model}
\alias{LEM}
\title{Concentration Summary Statistics to Estimated Exposure Curves.}
\description{
A function which draws exposure curves from a dataframe of summary statistic concentration data,
a selected weight column, an exposure factor, and n number of points.

This model is based on methods described in Egeghy & Lorber:
              "An assessment of the exposure of Americans to perfluorooctane
              sulfonate: a comparison of estimated intake with values inferred from
              NHANES data"(https://pubmed.ncbi.nlm.nih.gov/20145679/) and Lorber & Egeghy:
              Simple intake and pharmacokinetic modeling to characterize exposure of Americans to
              perfluoroctanoic acid, PFOA(https://pubmed.ncbi.nlm.nih.gov/21517063/)

The methods described above have been modified to include a geometric mean (GM) and
geometric standard deviation (GSD) estimating sequence from Pleil et. al (https://pubmed.ncbi.nlm.nih.gov/25333994/).

Accepted Units: "ng/m3", "ug/m3", "pg/ml", "ug/L", "ng/L", "pg/g", "ug/kg", "ng/g"


Concentration output: "ng/m3", "ng/L", or "ng/g"


WGMs and WGSDs are determined by complete cases of both, as visible in the 'Data' dataframe in the output list.

Function output returns five objects:

Description: The time and date of the run, the n number of points, the generated WGM and WGSD, , the colmun used to weight the results, and seed used.

Summary Statistics: the result, mean, min, 10th percentile, median, 75th percentile, 95th percentile, max for concentration and exposure.

Exposure: The generated lognormal exposure curve with "n" number of points.

Concentration: The generated lognormal concentration curve with "n" number of points.

Data: The input data used (df), including estimated summary statistics.

}
\usage{
LEM(df,wt,expf,expunits,n,seed)

In events where only concentration curves are required, please enter the concentration unit and set expf to 1.
}
\arguments{

df = A dataframe containing the following columns:
"Units","Sample Size","Min","Max","Median","Mean","SD","GM","GSD","P10","P25","P75","P90","P95","P99"

wt = A character input for the name of the 'weight' column (often sample size)

expf = the fraction by which the concentration is translated into exposure.

n = the number of generated exposure outcomes generated by rlnorm().

myseed = a seed used to dictate the random number generation (see "set.seed" for more)


}

\examples{
# Assuming a user has assembled an input "df" as described above:

LEM(df, wt, expf, expunits, n, myseed)
LEM(df = data, wt = "Sample Size", expf = 1.5, n = 200, myseed = 12345)
}
